# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HindaviSwarajya Seva Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/posts:
    get:
      operationId: listMyPosts
      tags: [me]
      summary: List the current user's seva posts (all approval statuses)
      responses:
        "200":
          description: User's seva posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SevaPost"
        "401":
          description: Not authenticated
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List seva posts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [Food, Education, Health, Shelter, Other]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [recent, likes, impact, comments]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of seva posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SevaPost"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a seva post (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaPost"
        "401":
          description: Not authenticated
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a seva post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seva post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaPost"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update one of your own seva posts (auth required, author-only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated seva post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaPost"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden — not the author
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete one of your own seva posts (auth required, author-only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden — not the author
        "404":
          description: Not found
  /posts/{id}/like:
    post:
      operationId: toggleLike
      tags: [posts]
      summary: Toggle like on a post (auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SevaPost"
        "401":
          description: Not authenticated
  /posts/{id}/comments:
    post:
      operationId: addComment
      tags: [posts]
      summary: Add comment to a post (auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentBody"
      responses:
        "201":
          description: Added comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Not authenticated
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [stats]
      summary: Get top seva contributors
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /tags/trending:
    get:
      operationId: getTrendingTags
      tags: [tags]
      summary: Get trending hashtags
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Trending tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendingTag"
...
components:
  schemas:
...
    Rank:
      type: string
      enum:
        - Sevak
        - Mavla
        - Bargir
        - Shiledar
        - Naik
        - Havaldar
        - Jumledar
        - Hazari
        - Panch Hazari
        - Sardar
        - Killedar
        - Deshpande
        - Deshmukh
        - Subhedar
        - Ashta Pradhan
        - Sarnobat
        - Sar Senapati
    SevaCategory:
      type: string
      enum: [Food, Education, Health, Shelter, Other]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: string
        location:
          type: string
        rank:
          $ref: "#/components/schemas/Rank"
        totalHelped:
          type: integer
        mudra:
          type: integer
          description: Seva points — totalHelped × 10
        chhava:
          type: boolean
          description: Honorary rank awarded by admins; displayed above the normal rank
        followersCount:
          type: integer
        postsCount:
          type: integer
      required:
        - id
        - name
        - avatar
        - location
        - rank
        - totalHelped
        - mudra
        - chhava
        - followersCount
        - postsCount
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        content:
          type: string
        timestamp:
          type: string
      required:
        - id
        - postId
        - userId
        - userName
        - userAvatar
        - content
        - timestamp
    SevaPost:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        content:
          type: string
        category:
          $ref: "#/components/schemas/SevaCategory"
        helpedPeople:
          type: integer
        likes:
          type: integer
        likedBy:
          type: array
          items:
            type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        tags:
          type: array
          items:
            type: string
        image:
          type: string
          nullable: true
        timestamp:
          type: string
        location:
          type: string
          nullable: true
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
      required:
        - id
        - userId
        - user
        - content
        - category
        - helpedPeople
        - likes
        - likedBy
        - comments
        - tags
        - timestamp
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
        category:
          $ref: "#/components/schemas/SevaCategory"
        helpedPeople:
          type: integer
        tags:
          type: array
          items:
            type: string
        image:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - content
        - category
        - helpedPeople
        - tags
    AddCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        totalHelped:
          type: integer
        postCount:
          type: integer
        totalLikes:
          type: integer
      required:
        - rank
        - user
        - totalHelped
        - postCount
        - totalLikes
    TrendingTag:
      type: object
      properties:
        tag:
          type: string
        count:
          type: integer
      required:
        - tag
        - count
...
    DeleteResult:
      type: object
      properties:
        deleted:
          type: boolean
      required:
        - deleted
    UpdatePostBody:
      type: object
      properties:
        content:
          type: string
        category:
          $ref: "#/components/schemas/SevaCategory"
        helpedPeople:
          type: integer
        location:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
...